+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (@egomobile/api-messenger)
 
+## 0.2.0
+
+- add `template` setting prop to [ISendApiMessageOptions](https://egomobile.github.io/node-api-messenger/interfaces/ISendApiMessageOptions.html)
+- fix response type of [sendMessages()](https://egomobile.github.io/node-api-messenger/interfaces/IApiMessenger.html#sendMessages)
+
 ## 0.1.1
 
 - initial release